scala 3, initial commit

+++ src/main/scala/com/juliasoft/sensors/core/SensorResult.scala
@@ -3,7 +3,7 @@ package com.juliasoft.sensors.core
 case class SensorResult(
   validMeasurements: Long,
   failedMeasurements: Long,
-  min: Option[Int],
-  average: Option[Int],
-  max: Option[Int],
+  min: Option[Byte],
+  average: Option[Byte],
+  max: Option[Byte],
 )
